% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ThreeMapComparison.R
\docType{class}
\name{AgreementBudget-class}
\alias{AgreementBudget-class}
\title{Class AgreementBudget}
\description{
An S4 class for information about sources of agreement and disagreement
between three categorical raster maps.
}
\section{Slots}{

\describe{
\item{\code{tables}}{list of data.frames that depict the three dimensional table
described by Pontius et al. (2011) at different resolutions}

\item{\code{factors}}{numeric vector of aggregation factors}

\item{\code{maps}}{list of RasterStack objects containing land use maps at different
resolutions}

\item{\code{categories}}{numeric vector of land use categories}

\item{\code{labels}}{character vector corresponding to \code{categories}}

\item{\code{overall}}{data.frame containing the overall agreement budget}

\item{\code{category}}{list of data.frames showing the agreement budget for each
category}

\item{\code{transition}}{list of data.frames showing the agreement budget for all
possible transitions}
}}

