% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-CategoryLabel.R
\docType{class}
\name{CategoryLabel-class}
\alias{CategoryLabel-class}
\title{Virtual class CategoryLabel}
\description{
A virtual S4 class to represent information about categorical Raster* objects.
}
\section{Slots}{

\describe{
\item{\code{categories}}{numeric vector of land use categories}

\item{\code{labels}}{character vector corresponding to \code{categories}}
}}

