% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPredictiveModelInputData.R
\name{getPredictiveModelInputData}
\alias{getPredictiveModelInputData}
\title{Extract data to fit predictive models}
\usage{
getPredictiveModelInputData(obs, ef, cells, ...)
}
\arguments{
\item{obs}{an ObsLulcRasterStack object}

\item{ef}{an ExpVarRasterList object}

\item{cells}{index of cells to be extracted, which may be a
\code{SpatialPoints*} object or a numeric vector representing cell numbers
(see \code{raster::\link[raster]{extract}})}

\item{...}{additional arguments to \link{as.data.frame}}
}
\value{
A data.frame.
}
\description{
Extract a data.frame containing variables required for fitting predictive
models.
}
\examples{

## TODO
}
\seealso{
\code{\link[base]{as.data.frame}}, \code{\link{ObsLulcRasterStack}},
\code{\link{ExpVarRasterList}}, \code{\link{partition}}
}
