% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Model.R
\docType{class}
\name{Model-class}
\alias{Model-class}
\title{Virtual class Model}
\description{
A virtual S4 class to represent land use change models.
}
\section{Slots}{

\describe{
\item{\code{output}}{RasterStack containing simulated land use maps or NULL}
}}

