% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-NeighbRasterStack.R
\docType{class}
\name{NeighbRasterStack-class}
\alias{NeighbRasterStack-class}
\title{Class NeighbRasterStack}
\description{
An S4 class for neighbourhood maps.
}
\section{Slots}{

\describe{
\item{\code{filename}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{layers}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{title}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{extent}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{rotated}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{rotation}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{ncols}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{nrows}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{crs}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{history}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{z}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{calls}}{list containing each call to \code{raster::\link[raster]{focal}}}

\item{\code{categories}}{numeric vector of land use categories for which neighbourhood
maps exist}
}}

