% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-ObsLulcRasterStack.R
\docType{class}
\name{ObsLulcRasterStack-class}
\alias{ObsLulcRasterStack-class}
\title{Class ObsLulcRasterStack}
\description{
An S4 class for observed land use maps.
}
\section{Slots}{

\describe{
\item{\code{filename}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{layers}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{title}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{extent}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{rotated}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{rotation}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{ncols}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{nrows}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{crs}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{history}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{z}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{t}}{numeric vector with timesteps corresponding to each observed map}

\item{\code{categories}}{numeric vector of land use categories}

\item{\code{labels}}{character vector corresponding to \code{categories}}
}}

