% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-PerformanceList.R
\docType{class}
\name{PerformanceList-class}
\alias{PerformanceList-class}
\title{Class PerformanceList}
\description{
An S4 class that extends \code{ROCR::\link[ROCR]{performance-class}} to hold
the results of multiple model evaluations.
}
\section{Slots}{

\describe{
\item{\code{performance}}{list of ROCR performance objects. Each object is
calculated for the corresponding ROCR prediction object held in the
PredictionList object supplied to the constructor function}

\item{\code{auc}}{numeric vector containing the area under the curve for each
performance object}

\item{\code{categories}}{numeric vector of land use categories for which performance
objects were created}

\item{\code{labels}}{character vector with labels corresponding to \code{categories}}
}}

