% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PerformanceList.R
\name{PerformanceList}
\alias{PerformanceList}
\title{Create a PerformanceList object}
\usage{
PerformanceList(pred, measure, x.measure = "cutoff", ...)
}
\arguments{
\item{pred}{an object of class PredictionList}

\item{measure}{performance measure to use for the evaluation. See
\code{ROCR::\link[ROCR]{performance}}}

\item{x.measure}{a second performance measure. See
\code{ROCR::\link[ROCR]{performance}}}

\item{\dots}{additional arguments to \code{ROCR::\link[ROCR]{performance}}}
}
\value{
A \code{PerformanceList} object.
}
\description{
This function uses different measures to evaluate multiple
\code{ROCR::\link[ROCR]{prediction}} objects stored in a
\code{\link{PredictionList}} object.
}
\examples{
## see lulcc-package examples
}
\references{
Sing, T., Sander, O., Beerenwinkel, N., Lengauer, T. (2005).
ROCR: visualizing classifier performance in R. Bioinformatics
21(20):3940-3941.
}
\seealso{
\code{\link{performance}}, \code{\link{PredictionList}}
}

