% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pie}
\alias{pie}
\title{Land use change dataset for Plum Island Ecosystem}
\format{A list containing the following elements:
\describe{
  \item{lu_pie_1985}{RasterLayer showing land use in 1985 (forest, built,
  other)}
  \item{lu_pie_1991}{RasterLayer showing land use in 1991}
  \item{lu_pie_1999}{RasterLayer showing land use in 1999}
  \item{ef_001}{RasterLayer showing elevation}
  \item{ef_002}{RasterLayer showing slope}
  \item{ef_003}{RasterLayer showing distance to built land in 1985}
}}
\usage{
pie
}
\description{
Dataset containing land use maps for 1985, 1991 and 1999 and several
explanatory variables derived from Pontius and Parmentier (2014).
}
\examples{

data(pie)
}
\references{
Pontius Jr, R. G., & Parmentier, B. (2014). Recommendations for using the
relative operating characteristic (ROC). Landscape ecology, 29(3), 367-382.
}
\keyword{datasets}
