% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-PredictionList.R
\docType{class}
\name{PredictionList-class}
\alias{PredictionList-class}
\title{Class PredictionList}
\description{
An S4 class that extends \code{ROCR::\link[ROCR]{prediction-class}} to hold
the results of multiple model predictions.
}
\section{Slots}{

\describe{
\item{\code{prediction}}{a list of \code{ROCR::\link[ROCR]{prediction-class}} objects.
These objects are calculated for each statistical model in the
\code{PredictiveModelList} object supplied to the constructor function}

\item{\code{categories}}{numeric vector of land use categories for which
\code{prediction} objects were created}

\item{\code{labels}}{character vector with labels corresponding to \code{categories}}
}}

