% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-PredictiveModelList.R
\docType{class}
\name{PredictiveModelList-class}
\alias{PredictiveModelList-class}
\title{Class PredictiveModelList}
\description{
An S4 class to hold multiple mathematical models for different land use
categories belonging to the same map.
}
\section{Slots}{

\describe{
\item{\code{models}}{list of predictive models}

\item{\code{categories}}{numeric vector of land use categories}

\item{\code{labels}}{character vector with labels corresponding to \code{categories}}
}}

