% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,ObsLulcRasterStack-method}
\alias{summary,ExpVarRasterList-method}
\alias{summary,NeighbRasterStack-method}
\alias{summary,PredictiveModelList-method}
\alias{summary,Model-method}
\title{Summary}
\usage{
summary(object, ...)

\S4method{summary}{ObsLulcRasterStack}(object, ...)

\S4method{summary}{ExpVarRasterList}(object, ...)

\S4method{summary}{NeighbRasterStack}(object, ...)

\S4method{summary}{PredictiveModelList}(object, ...)

\S4method{summary}{Model}(object, ...)
}
\arguments{
\item{object}{an object belonging to one of the classes in \code{lulcc}}

\item{...}{additional arguments (none)}
}
\value{
A matrix, data.frame or list
}
\description{
Summarise lulcc objects containing Raster* data or predictive models
}
