% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,Model,ANY-method}
\alias{plot,ObsLulcRasterStack,ANY-method}
\alias{plot,ThreeMapComparison,ANY-method}
\alias{plot.Model}
\alias{plot.ObsLulcRasterStack}
\alias{plot.ThreeMapComparison}
\title{Plot method for objects based on Raster* data}
\usage{
\method{plot}{ObsLulcRasterStack}(x, y, ...)

\method{plot}{Model}(x, y, ...)

\method{plot}{ThreeMapComparison}(x, y, category, factors, ...)

\S4method{plot}{ObsLulcRasterStack,ANY}(x, y, ...)

\S4method{plot}{Model,ANY}(x, y, ...)

\S4method{plot}{ThreeMapComparison,ANY}(x, y, category, factors, ...)
}
\arguments{
\item{x}{an object from \code{lulcc} containing Raster data}

\item{y}{not used}

\item{category}{numeric}

\item{factors}{numeric}

\item{\dots}{additional arguments to
  \code{rasterVis::\link[rasterVis]{levelplot}}}
}
\value{
A trellis object.
}
\description{
Plot \code{lulcc} objects based on Raster* data
}
\examples{
## see lulcc-package examples
}
\seealso{
\code{rasterVis::\link[rasterVis]{levelplot}}
}

